Adding all the questions I practiced in GE103
In this course, I have practiced 60 questions through an assignment named "Oceanverse", given by our professor S. R. S. Iyengar, HOD CSE, IIT Ropar. I have learnt a lot of basic and advanced-level concepts through these questions.

Questions which are in multiples of 10 are usually in the hard category.
Questions that are in multiples of 5 are usually in the medium category.
Other questions are generally of easy level - doable on the first try (though their level increases once we cross the 30-question mark).